% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_G.r
\name{Create_G}
\alias{Create_G}
\title{Generate Growth Matrix}
\usage{
Create_G(m, r, lambda, K, t, plot = FALSE)
}
\arguments{
\item{m}{number of spatial locations}

\item{r}{growth rate parameter}

\item{lambda}{intensity matrix (m x nT)}

\item{K}{carrying capacity (scalar)}

\item{t}{time point}

\item{plot}{(Default: FALSE)}
}
\description{
Computes the growth matrix, G. This represents a diagonal
matrix indexed by t.
}
