% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_M.r
\name{Create_M}
\alias{Create_M}
\title{Generate Movement Matrix}
\usage{
Create_M(m, longitude, latitude, tau, plot = FALSE)
}
\arguments{
\item{m}{number of spatial locations}

\item{longitude}{longitude of location}

\item{latitude}{latitude of location}

\item{tau}{dispersal parameter (vector of length m)}

\item{plot}{(Default: FALSE)}
}
\description{
Computes the movement matrix, M, based on a spatially varying Gaussian kernel
}
