% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleInvasion.r
\name{ExampleInvasion}
\alias{ExampleInvasion}
\title{Example Invasion Script}
\usage{
ExampleInvasion(m = 6, nT = 10, niter = 100, tune = list(h_lambda = 0.1,
  h_tau = 0.3, h_K = 20, h_r = 0.1))
}
\arguments{
\item{m}{number of spatial locations}

\item{nT}{number of time points}

\item{niter}{number of iterations}

\item{tune}{tuning parameters}
}
\description{
Example Invasion Script
}
