% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HootenMM.r
\name{HootenMM}
\alias{HootenMM}
\title{Hooten Space-Time Matrix Model}
\usage{
HootenMM(niter, m, nT, r, lambda, K, coords, tau, N, acceptR, Hparams,
  prior_lambda, prior_tau, prior_K, prior_r)
}
\arguments{
\item{niter}{number of iterations}

\item{m}{number of spatial locations}

\item{nT}{number of time points}

\item{r}{growth rate (scalar)}

\item{lambda}{intensity matrix (m x nT)}

\item{K}{carrying capacity (scalar)}

\item{coords}{matrix of spatial co-ordinates of locations (longitude, latitude)}

\item{tau}{dispersal parameter (vector of length m)}

\item{N}{true number of organisms (m x nT)}

\item{acceptR}{acceptance rates of parameters for M-H step}

\item{Hparams}{hyper-parameters for prior probability distributions}

\item{prior_lambda}{prior for lambda (mu_lambda, Sigma_lambda)}

\item{prior_tau}{prior for dispersal parameter, tau}

\item{prior_K}{prior for carrying capacity, K}

\item{prior_r}{prior for growth rate, r}
}
\description{
Uses the M-H algorithm to implement a BHM S-T model for invasions
}
\details{
This implements the M-H algorithm outlined in Hooten et al. (2007).
}
\references{
Hooten, M.B., Wikle, C.K., Dorazio, R.M. and Royle, J.A. (2007) Hierarchical Spatiotemporal Matrix
Models for Characterizing Invasions, Biometrics, 63, 558-567.
}
