% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.r
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{pass plot}

\item{plotlist}{plotlist}

\item{file}{file name}

\item{cols}{number of columns in layout}

\item{layout}{a matrix specifying the layout. If present, 'cols' is ignored.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.}
}
\description{
ggplot objects can be passed in ..., or to plotlist
(as a list of ggplot objects)
}
