% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunScenario.r
\name{RunScenario}
\alias{RunScenario}
\title{RunScenario}
\usage{
RunScenario(niter = 30000, nburn = 10000, nT = 10)
}
\arguments{
\item{niter}{Number of iterations}

\item{nburn}{Number of burn-ins}

\item{nT}{Number of time points}
}
\description{
Just a simple scenario to run Hooten Model
}
