% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleK.r
\name{Sample_K}
\alias{Sample_K}
\title{Sample the Carrying Capacity, K}
\usage{
Sample_K(tau, j, m, nT, r, lambda, K, coords, N, h_K, h_lambda, alpha_K,
  beta_K)
}
\arguments{
\item{tau}{dispersal parameter (vector of length m)}

\item{j}{MC iteration step}

\item{m}{number of spatial locations}

\item{nT}{number of time points}

\item{r}{growth rate (scalar)}

\item{lambda}{intensity matrix (m x nT)}

\item{K}{carrying capacity (scalar)}

\item{coords}{matrix of spatial co-ordinates of locations (longitude, latitude)}

\item{N}{true number of organisms (m x nT)}

\item{h_K}{M-H step value for K}

\item{h_lambda}{M-H step value for lambda}

\item{alpha_K}{prior parameter for K}

\item{beta_K}{prior parameter for K}
}
\description{
Simulates from the posterior distribution of the carrying capacity parameter, K.
}
