% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleLambda.r
\name{Sample_lambda}
\alias{Sample_lambda}
\title{Sample the intensity, lambda}
\usage{
Sample_lambda(m, nT, r, j, lambda, K, coords, tau, N, mu_lambda, sigma_lambda,
  h_lambda)
}
\arguments{
\item{m}{number of spatial locations}

\item{nT}{number of time points}

\item{r}{growth rate (scalar)}

\item{j}{MC iteration step}

\item{lambda}{intensity matrix (m x nT)}

\item{K}{carrying capacity (scalar)}

\item{coords}{matrix of spatial co-ordinates of locations (longitude, latitude)}

\item{tau}{dispersal parameter (vector of length m)}

\item{N}{true number of organisms (m x nT)}

\item{mu_lambda}{prior mean vector for lambda}

\item{sigma_lambda}{prior variance matrix for lambda}

\item{h_lambda}{M-H step value for lambda}
}
\description{
Simulates from the posterior distribution of the intensity, lambda.
}
