% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_r.r
\name{Sample_r}
\alias{Sample_r}
\title{Sample Growth Rate, r}
\usage{
Sample_r(tau, j, m, nT, r, lambda, K, coords, N, h_r, h_lambda, mu_r, sig2_r)
}
\arguments{
\item{tau}{dispersal parameter (vector of length m)}

\item{j}{MC iteration step}

\item{m}{number of spatial locations}

\item{nT}{number of time points}

\item{r}{growth rate (scalar)}

\item{lambda}{intensity matrix (m x nT)}

\item{K}{carrying capacity (scalar)}

\item{coords}{matrix of spatial co-ordinates of locations (longitude, latitude)}

\item{N}{true number of organisms (m x nT)}

\item{h_r}{M-H step value for r}

\item{h_lambda}{M-H step value for lambda}

\item{mu_r}{prior mean for r}

\item{sig2_r}{prior variance for r}
}
\description{
Simulates from the posterior distribution of the growth rate, r.
}
