% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleTau.r
\name{Sample_tau}
\alias{Sample_tau}
\title{Sample the dispersal parameter, tau}
\usage{
Sample_tau(tau, j, m, nT, r, lambda, K, coords, N, h_tau, h_lambda, mu_tau,
  sigma_tau)
}
\arguments{
\item{tau}{dispersal parameter (vector of length m)}

\item{j}{MC iteration step}

\item{m}{number of spatial locations}

\item{nT}{number of time points}

\item{r}{growth rate (scalar)}

\item{lambda}{intensity matrix (m x nT)}

\item{K}{carrying capacity (scalar)}

\item{coords}{matrix of spatial co-ordinates of locations (longitude, latitude)}

\item{N}{true number of organisms (m x nT)}

\item{h_tau}{M-H step value for tau}

\item{h_lambda}{M-H step value for lambda}

\item{mu_tau}{prior mean for tau}

\item{sigma_tau}{prior variance matrix for tau}
}
\description{
Simulates from the posterior distribution of the dispersal parameter, tau.
}
