% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupHSTMM.r
\name{SetupHSTMM}
\alias{SetupHSTMM}
\title{Set up the HSTMM}
\usage{
SetupHSTMM(nT = 10, niter = 100, coords, m = nrow(coords),
  tune = list(h_lambda = 0.1, h_tau = 0.3, h_K = 20, h_r = 0.1),
  prior_lambda = list(lambda_m = log(15), lambda_sig = log(1.1)),
  prior_tau = list(tau_m = log(2), tau_sig = log(1.1)), prior_K = list(alpha
  = 40, beta = 50), prior_r = list(mu = 0.5, sig2 = 0.25), initialV = list(K
  = 150, r = 0.3))
}
\arguments{
\item{nT}{number of time steps}

\item{niter}{number of MCMC iterations}

\item{coords}{matrix of longitude and latitude co-ordinates}

\item{m}{number of spatial locations}

\item{tune}{tuning parameters for MH step. A value is listed for each type of parameter.}

\item{prior_lambda}{prior for lambda}

\item{prior_tau}{prior for tau}

\item{prior_K}{prior for K}

\item{prior_r}{prior for r}

\item{initialV}{initial values}
}
\description{
Set up the HSTMM
}
