% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleInvasion.r
\name{ExampleInvasion}
\alias{ExampleInvasion}
\title{Example Invasion Script}
\usage{
ExampleInvasion(m = 6, nT = 10, niter = 100, coords,
  prior_lambda = prior_lambda, prior_tau = prior_tau,
  prior_K = prior_K, prior_r = prior_r, prior_theta,
  initialV = initialV, tune = list(h_lambda = 0.1, h_tau = 0.3, h_K =
  20, h_r = 0.1, h_N = 1), sample_n)
}
\arguments{
\item{m}{number of spatial locations}

\item{nT}{number of time points}

\item{niter}{number of iterations}

\item{coords}{coordinates}

\item{prior_lambda}{prior for lambda}

\item{prior_tau}{prior for tau}

\item{prior_K}{prior for K}

\item{prior_r}{prior for r}

\item{prior_theta}{prior for theta}

\item{initialV}{initialV}

\item{tune}{tuning parameters}

\item{sample_n}{sample_n}
}
\description{
Example Invasion Script
}
